requirements.txt
aiohttp==3.8.1
asyncer==0.0.2
click==8.1.3
fastapi==0.92.0
filetype==1.2.0
imagehash==4.3.1
numpy==1.23.5
onnxruntime==1.14.1
opencv-python-headless==4.6.0.66
pillow==9.3.0
pooch==1.6.0
pymatting==1.1.8
python-multipart==0.0.5
scikit-image==0.19.3
scipy==1.9.3
tqdm==4.64.1
uvicorn==0.20.0
watchdog==2.1.9
